requirements.txt
tensorflow==2.12.0
Flask
numpy
requests
pillow
opencv-python